requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.7
django-extensions==3.2.3
django-filter==23.5
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
executing==2.0.1
flake8==6.1.0
graphviz==0.20.1
idna==3.6
inflection==0.5.1
ipython==8.19.0
isort==5.13.2
jedi==0.19.1
Markdown==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydot==2.0.0
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
